Week-4/presidio-agent/requirements.txt
mcp
google-auth
google-auth-oauthlib
google-api-python-client
langchain-chroma
langchain-aws
langchain-core
python-dotenv
langchain-tavily
langchain-community
langchain-text-splitters
asyncio
